requirements.txt
beautifulsoup4
requests
tqdm
python-dotenv
pyyaml
PyPDF2
sentence-transformers
chromadb
openai
